fix: reduce fields fetched by dashboards query to that which is needed
