+++ src/components/Dashboard/Dashboard.js
@@ -87,7 +87,7 @@ const Dashboard = ({
 
     useEffect(() => {
         if (dashboardsLoaded && !dashboardsIsEmpty) {
-            selectDashboard(routeId)
+            selectDashboard(routeId, mode)
         }
     }, [dashboardsLoaded, dashboardsIsEmpty, routeId, mode])
